requirements.txt
streamlit==1.28.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.7
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-cohere==0.3.2
cohere==5.5.8
